+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file. Starting wi
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+  * `Contract`; shortcode:  `con`
+  * `LineItem`; shortcode:  `li`
+  * `LineItemEvent`; shortcode:  `lie`
+  * `OrderEvent`; shortcode:  `oe`
+  * `Term`; shortcode:  `term`
+  * `QueuedSubscriptionRequest`; shortcode: `qsr`
+  * `Webhook`; shorcode: `wh`,
+  * `QsrService`; shortcode: `qsrsrvc`
+  * `QsrCache`; shortcode `QsrCache` (shortcode is capitalized to indicate it is a special purpose Oid for managing local caches)
+### Removed
+### Changed
+### Fixed
+### Deprecated
+### Security
+
+
 ## [4.0.7] -- 2020-09-08
 
 ### Added